+++ tests/testthat/setup.R
@@ -1,0 +1,3 @@
+op <- options(getRad.progress = FALSE)
+
+withr::defer(options(op), teardown_env())