+++ NEWS.md
@@ -1,6 +1,7 @@
 # getRad (development version)
 
 * Use options for urls so they can be changes without a package update (#120).
+* Add `getRad.progress` option to silence progress.
 
 # getRad 0.3.0